chatbot/requirements.txt
langchain
langchain-experimental
llamaapi
openai
huggingface_hub
python-dotenv
streamlit